+++ lib/auth0/logout_user/logout_user_mobile.dart
@@ -5,6 +5,7 @@ import 'package:auth0_flutter2/auth0/init_auth0/init_auth0_mobile.dart';
 Future<void> logoutUser({
   required String auth0Domain,
   required String auth0ClientId,
+  required String redirectUri,
   String? scheme,
 }) async {
   Auth0 auth0 = initAuth0Mobile(
@@ -13,7 +14,9 @@ Future<void> logoutUser({
   );
 
   try {
-    await auth0.webAuthentication(scheme: scheme).logout();
+    scheme != null
+        ? await auth0.webAuthentication(scheme: scheme).logout()
+        : await auth0.webAuthentication().logout();
   } catch (e) {
     throw Exception(e);
   }