+++ lib/auth0/logout_user/logout_user_stub.dart
@@ -2,6 +2,7 @@
 Future<void> logoutUser({
   required String auth0Domain,
   required String auth0ClientId,
+  required String redirectUri,
   String? scheme,
 }) =>
     throw UnsupportedError(