+++ lib/auth0/logout_user/logout_user_web.dart
@@ -1,3 +1,4 @@
+import 'package:auth0_flutter2/auth0/auth0_flutter_web/src/options.dart';
 import 'package:auth0_flutter2/auth0/init_auth0/init_auth0_web.dart';
 import 'package:auth0_flutter2/auth0/auth0_flutter_web/auth0_flutter_web.dart'
     as auth0_web;
@@ -6,15 +7,21 @@ import 'package:auth0_flutter2/auth0/auth0_flutter_web/auth0_flutter_web.dart'
 Future<void> logoutUser({
   required String auth0Domain,
   required String auth0ClientId,
+  required String redirectUri,
   String? scheme,
 }) async {
   auth0_web.Auth0 auth0 = await initAuth0Web(
     auth0Domain: auth0Domain,
     auth0ClientId: auth0ClientId,
+    redirectUri: redirectUri,
   );
 
   try {
-    auth0.logout();
+    auth0.logout(
+      options: LogoutOptions(
+        returnTo: redirectUri,
+      ),
+    );
   } catch (e) {
     throw Exception(e);
   }